requirements.txt
numpy==1.26.4
scikit-learn==1.3.2
sentence-transformers==2.2.2
spacy==3.7.2
pdfminer.six==20221105
python-docx==1.1.0
requests==2.31.0
beautifulsoup4==4.12.2
pydantic==2.5.3
openai==1.3.7
pytest==7.4.4
